refactoring payment page
